@@ -11,10 +11,13 @@
     "axios": "^1.6.7",
     "nanoid": "^5.0.4",
     "prop-types": "^15.8.1",
-    "react": "^18.1.0",
-    "react-dom": "^18.1.0",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "react-helmet": "^6.1.0",
     "react-redux": "^9.1.0",
+    "react-router-dom": "^6.22.2",
     "react-scripts": "5.0.1",
+    "redux-persist": "^6.0.0",
     "web-vitals": "^2.1.3"
   },
   "scripts": {
