@@ -1,6 +1,6 @@
 import { useState } from 'react';
 import { useDispatch } from 'react-redux';
-import { addContact } from '../redux/operations';
+import { addContact } from '../redux/contacts/operations';
 
 const ContactForm = () => {
   const [name, setName] = useState('');
